Decode salsify_uuid_binary in text format [PA-1346] (#6)
The salsify_uuid_binary schema stores UUIDs as 16 bytes to save on
space. However, we want to convert those values to their text form when
storing in Elasticsearch.

Avro supports logical types which perform type conversions on
read. Unfortunately we don't have a logical type on salsify_uuid_binary,
however we can tie into the type adapter with our own custom logic.

The salsify_uuid_binary is defined as a "fixed" type in Avro, so this
change injects a patch to override the adapter logic for that type. The
encoding part will accept either the binary or text representation when
writing. When reading it should always be in the 16 byte format.

This library didn't have a functional test suite. Most of the changes
are getting a deserialization test working. We pull in avro_turf in the
test suite for a reference implementation of using the schema registry.